Fix metal shader export and extract common shader bundle loading into BackgroundHelper
