+++ Assets/Script/Gameplay/BackgroundManager.cs
@@ -17,7 +17,7 @@ using YARG.Settings;
 using YARG.Venue;
 using YARG.Venue.Characters;
 using YARG.Core.Logging;
-using ExportType = YARG.Venue.BundleBackgroundManager.ExportType;
+using YARG.Helpers;
 
 #if UNITY_EDITOR
 using UnityEditor.SceneManagement;
@@ -74,9 +74,9 @@ namespace YARG.Gameplay
         private BundleBackgroundManager _bundleBackgroundManager;
 
 #if UNITY_EDITOR
-        private bool        _usingEditorVenue;
-        private string      _editorVenuePath;
-        private Scene       _editorVenueScene;
+        private          bool             _usingEditorVenue;
+        private          string           _editorVenuePath;
+        private          Scene            _editorVenueScene;
 #endif
         // "The Unity message 'Start' has an incorrect signature."
         [SuppressMessage("Type Safety", "UNT0006", Justification = "UniTaskVoid is a compatible return type.")]
@@ -203,7 +203,7 @@ namespace YARG.Gameplay
             var renderers = bg.GetComponentsInChildren<Renderer>(true);
 
             // Load Metal shaders, if necessary
-            shaderBundle = await LoadMetalShaders(bundle, bg, ExportType.Background);
+            shaderBundle = await BackgroundHelper.LoadMetalShaders(bundle, bg, BackgroundHelper.ExportType.Background);
 
             // Hookup song-specific textures
             var textureManager = GetComponent<TextureManager>();
@@ -515,7 +515,7 @@ namespace YARG.Gameplay
             }
 
             // Load Metal shaders
-            var shaderBundle = await LoadMetalShaders(bundle, character, ExportType.Character);
+            var shaderBundle = await BackgroundHelper.LoadMetalShaders(bundle, character, BackgroundHelper.ExportType.Character);
             if (shaderBundle != null)
             {
                 _bundleBackgroundManager.ShaderBundles.Add(shaderBundle);
@@ -577,76 +577,6 @@ namespace YARG.Gameplay
         }
 
 #pragma warning disable CS1998 // Async method lacks 'await' operators and will run synchronously
-        private async UniTask<AssetBundle> LoadMetalShaders(AssetBundle bundle, GameObject bg, ExportType type)
-#pragma warning restore CS1998 // Async method lacks 'await' operators and will run synchronously
-        {
-#if UNITY_EDITOR_OSX || UNITY_STANDALONE_OSX
-            AssetBundle shaderBundle = null;
-            var renderers = bg.GetComponentsInChildren<Renderer>(true);
-            var metalShaders = new Dictionary<string, Shader>();
-
-            var shaderBundleName = type switch
-            {
-                ExportType.Character => "Assets/" + BundleBackgroundManager.CHARACTER_SHADER_BUNDLE_NAME,
-                ExportType.Background => "Assets/" + BundleBackgroundManager.BACKGROUND_SHADER_BUNDLE_NAME,
-                _ => throw new ArgumentOutOfRangeException(nameof(type), type, null)
-            };
-
-            var shaderBundleData = (TextAsset)await bundle.LoadAssetAsync<TextAsset>(
-                shaderBundleName
-            );
-
-            if (shaderBundleData == null)
-            {
-                shaderBundleData = (TextAsset)await bundle.LoadAssetAsync<TextAsset>(
-                    bundle.name + BundleBackgroundManager.BUNDLE_OSX_SUFFIX
-                );
-            }
-
-            if (shaderBundleData != null && shaderBundleData.bytes.Length > 0)
-            {
-                YargLogger.LogInfo("Loading Metal shader bundle");
-                shaderBundle = await AssetBundle.LoadFromMemoryAsync(shaderBundleData.bytes);
-                var allAssets = shaderBundle.LoadAllAssets<Shader>();
-                foreach (var shader in allAssets)
-                {
-                    metalShaders.Add(shader.name, shader);
-                }
-            }
-            else
-            {
-                YargLogger.LogInfo("Did not find Metal shader bundle");
-            }
-
-            // Yarground comes with shaders for dx11/dx12/glcore/vulkan
-            // Metal shaders used on OSX come in this separate bundle
-            // Update our renderers to use them
-
-            foreach (var renderer in renderers)
-            {
-                foreach (var material in renderer.sharedMaterials)
-                {
-                    var shaderName = material.shader.name;
-                    if (metalShaders.TryGetValue(shaderName, out var shader))
-                    {
-                        YargLogger.LogFormatDebug("Found bundled shader {0}", shaderName);
-                        // We found shader from Yarground
-                        material.shader = shader;
-                    }
-                    else
-                    {
-                        YargLogger.LogFormatDebug("Did not find bundled shader {0}", shaderName);
-                        // Fallback to try to find among builtin shaders
-                        material.shader = Shader.Find(shaderName);
-                    }
-                }
-            }
-
-            return shaderBundle;
-#endif
-            // Fallback if we're not running on OSX
-            return null;
-        }
 
         // It would be better if we could replace all references, but I'm not sure how to do that, so I'm fixing up the ones I know how to do
         public void ReplaceReferences(GameObject venueRoot, GameObject oldObject, GameObject newObject)
