+++ Assets/Script/Settings/Metadata/PreviewBuilders/CharacterPreviewBuilder.cs
@@ -5,6 +5,7 @@ using UnityEngine;
 using UnityEngine.AddressableAssets;
 using UnityEngine.UI;
 using YARG.Core.Logging;
+using YARG.Gameplay;
 using YARG.Helpers;
 using YARG.Helpers.Extensions;
 using YARG.Menu.Settings;
@@ -226,60 +227,7 @@ namespace YARG.Settings.Metadata
         private static async UniTask<AssetBundle> LoadMetalShaders(AssetBundle bundle, GameObject bg)
 #pragma warning restore CS1998 // Async method lacks 'await' operators and will run synchronously
         {
-#if UNITY_EDITOR_OSX || UNITY_STANDALONE_OSX
-            AssetBundle shaderBundle = null;
-            var renderers = bg.GetComponentsInChildren<Renderer>(true);
-            var metalShaders = new Dictionary<string, Shader>();
-
-            var shaderBundleName = "Assets/" + BundleBackgroundManager.CHARACTER_SHADER_BUNDLE_NAME;
-
-            var shaderBundleData = (TextAsset)await bundle.LoadAssetAsync<TextAsset>(
-                shaderBundleName
-            );
-
-            if (shaderBundleData != null && shaderBundleData.bytes.Length > 0)
-            {
-                YargLogger.LogInfo("Loading Metal shader bundle");
-                shaderBundle = await AssetBundle.LoadFromMemoryAsync(shaderBundleData.bytes);
-                var allAssets = shaderBundle.LoadAllAssets<Shader>();
-                foreach (var shader in allAssets)
-                {
-                    metalShaders.Add(shader.name, shader);
-                }
-            }
-            else
-            {
-                YargLogger.LogInfo("Did not find Metal shader bundle");
-            }
-
-            // Yarground comes with shaders for dx11/dx12/glcore/vulkan
-            // Metal shaders used on OSX come in this separate bundle
-            // Update our renderers to use them
-
-            foreach (var renderer in renderers)
-            {
-                foreach (var material in renderer.sharedMaterials)
-                {
-                    var shaderName = material.shader.name;
-                    if (metalShaders.TryGetValue(shaderName, out var shader))
-                    {
-                        YargLogger.LogFormatDebug("Found bundled shader {0}", shaderName);
-                        // We found shader from Yarground
-                        material.shader = shader;
-                    }
-                    else
-                    {
-                        YargLogger.LogFormatDebug("Did not find bundled shader {0}", shaderName);
-                        // Fallback to try to find among builtin shaders
-                        material.shader = Shader.Find(shaderName);
-                    }
-                }
-            }
-
-            return shaderBundle;
-#endif
-            // Fallback if we're not running on OSX
-            return null;
+            return await BackgroundHelper.LoadMetalShaders(bundle, bg, BackgroundHelper.ExportType.Character);
         }
     }
 }
